requirements.txt
schedule==0.6.0
requests==2.22.0
beautifulsoup4==4.9.3
python-dotenv==0.15.0
